♻️ refactor(tools): rename package coretools to tools
Package name now matches directory name. Callers that also import
pkg/tools alias it as pkgtools to disambiguate.
