+++ internal/agent/coretools_builder.go
@@ -1,16 +1,16 @@
 package agent
 
 import (
-	coretools "github.com/CherryHQ/stella/internal/tools"
+	"github.com/CherryHQ/stella/internal/tools"
 	pkgsandbox "github.com/CherryHQ/stella/pkg/sandbox"
-	"github.com/CherryHQ/stella/pkg/tools"
+	pkgtools "github.com/CherryHQ/stella/pkg/tools"
 	plugintools "github.com/CherryHQ/stella/plugins/tools"
 )
 
 // buildSandboxCoreTools creates core tools using the active sandbox session.
-func buildSandboxCoreTools(session pkgsandbox.Session, bc plugintools.BuildContext) []tools.Tool {
+func buildSandboxCoreTools(session pkgsandbox.Session, bc plugintools.BuildContext) []pkgtools.Tool {
 	if session == nil {
 		return nil
 	}
-	return coretools.New(session, bc.Paths.ToolsBinDir, bc.Paths.ProjectRoot)
+	return tools.New(session, bc.Paths.ToolsBinDir, bc.Paths.ProjectRoot)
 }
