+++ server/tools.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"sort"
 
-	coretools "github.com/CherryHQ/stella/internal/tools"
+	"github.com/CherryHQ/stella/internal/tools"
 	delegatetool "github.com/CherryHQ/stella/internal/tools/delegate"
 	"github.com/CherryHQ/stella/pkg/memory"
 	pkgtools "github.com/CherryHQ/stella/pkg/tools"
@@ -28,29 +28,29 @@ func defToJSON(def pkgtools.Definition, category string) toolJSON {
 }
 
 func (s *Server) ListTools(w http.ResponseWriter, r *http.Request) {
-	var tools []toolJSON
+	var items []toolJSON
 
-	for _, def := range coretools.Definitions() {
-		tools = append(tools, defToJSON(def, "builtin"))
+	for _, def := range tools.Definitions() {
+		items = append(items, defToJSON(def, "builtin"))
 	}
 
 	// Delegate tool (always present).
-	tools = append(tools, defToJSON(delegatetool.DelegateDefinition(nil), "builtin"))
+	items = append(items, defToJSON(delegatetool.DelegateDefinition(nil), "builtin"))
 
 	// Builtin tools (memory, skills).
 	for _, def := range s.builtinToolDefinitions() {
-		tools = append(tools, defToJSON(def, "builtin"))
+		items = append(items, defToJSON(def, "builtin"))
 	}
 
-	sort.Slice(tools, func(i, j int) bool {
-		if tools[i].Category != tools[j].Category {
+	sort.Slice(items, func(i, j int) bool {
+		if items[i].Category != items[j].Category {
 			order := map[string]int{"builtin": 0, "extra": 1}
-			return order[tools[i].Category] < order[tools[j].Category]
+			return order[items[i].Category] < order[items[j].Category]
 		}
-		return tools[i].Name < tools[j].Name
+		return items[i].Name < items[j].Name
 	})
 
-	writeData(w, http.StatusOK, tools)
+	writeData(w, http.StatusOK, items)
 }
 
 // builtinToolDefinitions returns the canonical definitions from each tool
